+++ accounts/views.py
@@ -1,7 +1,7 @@
-from django.shortcuts import render
 from allauth.socialaccount.models import SocialAccount, SocialToken
-from django.contrib.auth.decorators import login_required
 from django.contrib.auth import get_user_model
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import render
 
 User = get_user_model()